When navigating using AJAX, push state to browser history so you can use browser back/forward buttons
To get those buttons to work, we force the browser to load the page fully if back/forward has got us to a page that was put in the browser history by pushState. Drawbacks are that this causes a full page load, and it doesn't set the referer properly, so coming back from another section of the site doesn't animate the nav as it should.